IGNITE-24481 Move constants to private
